Should Jetty be able to serve JSP with Java custom tag and TLD in WEB-INF/tags ?
