+++ src/main/resources/webroot/WEB-INF/bye-from-web-inf.tld
@@ -1,0 +1,14 @@
+<!DOCTYPE taglib
+        PUBLIC "-//Sun Microsystems, Inc.//DTD JSP Tag Library 1.2//EN"
+        "http://java.sun.com/dtd/web-jsptaglibrary_1_2.dtd">
+
+<taglib>
+    <tlib-version>1.0</tlib-version>
+    <jsp-version>1.2</jsp-version>
+    <short-name>Leaving a lot behind</short-name>
+
+    <tag>
+        <name>scripting</name>
+        <tag-class>org.eclipse.jetty.demo.tags.Leaving</tag-class>
+    </tag>
+</taglib>
