+++ src/main/resources/webroot/test/tagclass-with-tld-in-web-inf-tags.jsp
@@ -1,0 +1,7 @@
+<%@ taglib uri="/WEB-INF/tags/bye-from-web-inf-tags.tld" prefix="leave" %>
+
+<html>
+<body>
+    <leave:scripting />
+</body>
+</html>
