+++ src/main/resources/webroot/test/tagclass-with-tld-in-web-inf.jsp
@@ -1,0 +1,7 @@
+<%@ taglib uri="/WEB-INF/bye-from-web-inf.tld" prefix="leave" %>
+
+<html>
+<body>
+    <leave:scripting />
+</body>
+</html>
